+++ tests/dag_tests.sh
@@ -236,9 +236,14 @@ if [ $1 = "atac-seq" ]; then
   seq2science run atac-seq -n --cores $CORES --configfile tests/$WF/genrich_macs2.yaml --snakemakeOptions dryrun=True quiet=True config={samples:tests/atac_seq/complex_samples.tsv,create_qc_report:True} | tee tests/local_test_results/${1}_dag
   assert_rulecount $1 featureCounts 16
 
-  printf "\control and merging of tecnical replicates\n"
+  printf "\ncontrol and merging of tecnical replicates\n"
   seq2science run atac-seq -n --cores $CORES --configfile tests/$WF/genrich_macs2.yaml --snakemakeOptions dryrun=True quiet=True config={samples:tests/atac_seq/control.tsv,create_qc_report:True} | tee tests/local_test_results/${1}_dag
   assert_rulecount $1 bwa_mem 7
+
+  printf "\ninput control different across same condition\n"
+  seq2science run atac-seq -n --cores $CORES --configfile tests/$WF/genrich_macs2.yaml --snakemakeOptions dryrun=True quiet=True config={samples:tests/atac_seq/complex_samples2.tsv,create_qc_report:True} | tee tests/local_test_results/${1}_dag
+  assert_rulecount $1 genrich_pileup 4
+  assert_rulecount $1 macs2_callpeak 4
 
   test_ran=1
 fi